requirements.txt
fastapi
uvicorn
firebase-admin
pydantic
python-dotenv
openai
starlette
google-cloud-firestore